# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Committee Management Service
    version: 0.0.1
host: localhost:80
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /committees:
        post:
            tags:
                - committee-service
            summary: create-committee committee-service
            description: Create Committee
            operationId: committee-service#create-committee
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: false
                  type: string
                  enum:
                    - "1"
                - name: Authorization
                  in: header
                  description: JWT token issued by Heimdall
                  required: false
                  type: string
                - name: Create-CommitteeRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CommitteeServiceCreateCommitteeRequestBody'
                    required:
                        - name
                        - category
                        - project_uid
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/CommitteeFullWithReadonlyAttributes'
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - message
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - message
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ConflictError'
                        required:
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /committees/{uid}:
        get:
            tags:
                - committee-service
            summary: get-committee-base committee-service
            description: Get Committee
            operationId: committee-service#get-committee-base
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: false
                  type: string
                  enum:
                    - "1"
                - name: uid
                  in: path
                  description: Committee UID -- v2 uid, not related to v1 id directly
                  required: true
                  type: string
                  format: uuid
                - name: Authorization
                  in: header
                  description: JWT token issued by Heimdall
                  required: false
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CommitteeServiceGetCommitteeBaseResponseBody'
                    headers:
                        ETag:
                            description: ETag header value
                            type: string
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
        put:
            tags:
                - committee-service
            summary: update-committee-base committee-service
            description: Update Committee
            operationId: committee-service#update-committee-base
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: false
                  type: string
                  enum:
                    - "1"
                - name: uid
                  in: path
                  description: Committee UID -- v2 uid, not related to v1 id directly
                  required: true
                  type: string
                  format: uuid
                - name: Authorization
                  in: header
                  description: JWT token issued by Heimdall
                  required: false
                  type: string
                - name: If-Match
                  in: header
                  description: If-Match header value for conditional requests
                  required: false
                  type: string
                - name: Update-Committee-BaseRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CommitteeServiceUpdateCommitteeBaseRequestBody'
                    required:
                        - name
                        - category
                        - project_uid
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CommitteeBaseWithReadonlyAttributes'
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - message
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - message
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ConflictError'
                        required:
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
        delete:
            tags:
                - committee-service
            summary: delete-committee committee-service
            description: Delete Committee
            operationId: committee-service#delete-committee
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: false
                  type: string
                  enum:
                    - "1"
                - name: uid
                  in: path
                  description: Committee UID -- v2 uid, not related to v1 id directly
                  required: true
                  type: string
                  format: uuid
                - name: Authorization
                  in: header
                  description: JWT token issued by Heimdall
                  required: false
                  type: string
                - name: If-Match
                  in: header
                  description: If-Match header value for conditional requests
                  required: false
                  type: string
            responses:
                "204":
                    description: No Content response.
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - message
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - message
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ConflictError'
                        required:
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /committees/{uid}/members:
        post:
            tags:
                - committee-members-service
            summary: create-committee-member committee-members-service
            description: Add a new member to a committee
            operationId: committee-members-service#create-committee-member
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: true
                  type: string
                  enum:
                    - "1"
                - name: uid
                  in: path
                  description: Committee UID -- v2 uid, not related to v1 id directly
                  required: true
                  type: string
                  format: uuid
                - name: Authorization
                  in: header
                  description: JWT token issued by Heimdall
                  required: false
                  type: string
                - name: Create-Committee-MemberRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CommitteeMembersServiceCreateCommitteeMemberRequestBody'
                    required:
                        - email
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/CommitteeMemberFullWithReadonlyAttributes'
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - message
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - message
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/ConflictError'
                        required:
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
...
definitions:
    BadRequestError:
        title: BadRequestError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: The request was invalid.
        description: Bad request
        example:
            message: The request was invalid.
        required:
            - message
...
    CommitteeServiceGetCommitteeBaseResponseBody:
        title: CommitteeServiceGetCommitteeBaseResponseBody
        $ref: '#/definitions/CommitteeBaseWithReadonlyAttributes'
...
    ConflictError:
        title: ConflictError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: The resource already exists.
        description: Conflict
        example:
            message: The resource already exists.
        required:
            - message
    InternalServerError:
        title: InternalServerError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: An internal server error occurred.
        description: Internal server error
        example:
            message: An internal server error occurred.
        required:
            - message
    NotFoundError:
        title: NotFoundError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: The resource was not found.
        description: Resource not found
        example:
            message: The resource was not found.
        required:
            - message
    ServiceUnavailableError:
        title: ServiceUnavailableError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: The service is unavailable.
        description: Service unavailable
        example:
            message: The service is unavailable.
        required:
            - message
securityDefinitions:
    jwt_header_Authorization:
        type: apiKey
        description: Heimdall authorization
        name: Authorization
        in: header